base: initialize zoslib using struct
It is possible that we may add more environment variables in the future.
We don't want to be required to update every `init_zoslib()` call site
when this happens. So we change the arguments used to initialize zoslib
into a single struct instead. This way, we can easily modify the struct
to add more members in the future without changing `init_zoslib()`.

PR-URL: https://github.ibm.com/open-z/zoslib/pull/13