backtrace: change format to match other platforms
Includes the frame number in the backtrace, and re-arranges
the other fields to match the format on other platforms such
as LoZ and linux.

Also adds the right parantheses if missing from the entry name,
i.e. 'foo(long list-of-args...' becomes 'foo(long list-of-args...)'
or 'foo(long list-of-args..)' if already at max length.