test: fix getentropy() buffer is not null check (#104)
There was a bug with getentropy() that caused the buffer to be null.
Previously, a test was introduced which checked to ensure that the first
element in the buffer was not 0. However, since the buffer is filled
with random values, 0 is a valid value for the first element. Instead,
the entire buffer should be checked to make sure that at least one
element is non-zero.