add support for envar __MEMORY_USAGE_LOG_INC
This tells zoslib to display memory allocation messages only when
the current allocated size increases by the given value (in bytes).
The intent is to reduce the size of the log file esp for long-running
apps, which for this to be enabled, the __MEMORY_USAGE_LOG_LEVEL
should be set to 0.